fix: add tests. to avoid race condition one token creation and token renewal
